Data Analysis with Python
Assignment:

You are a Data Analyst working at a Real Estate Investment Trust. The Trust would like to start investing in Residential real estate. You are tasked with determining the market price of a house given a set of features. You will analyze and predict housing prices using attributes or characteristics such as square footage, number of bedrooms, number of floors, and so on.